+++ src/test/java/com/github/m0nk3y2k4/thetvdb/internal/connection/RemoteAPITest.java
@@ -16,38 +16,25 @@
 
 package com.github.m0nk3y2k4.thetvdb.internal.connection;
 
+import static com.github.m0nk3y2k4.thetvdb.internal.connection.RemoteAPI.API_V4;
 import static com.github.m0nk3y2k4.thetvdb.internal.connection.RemoteAPI.HTTPS;
 import static com.github.m0nk3y2k4.thetvdb.internal.connection.RemoteAPI.HTTPS_DEFAULT;
 import static com.github.m0nk3y2k4.thetvdb.internal.connection.RemoteAPI.THE_TVDB_DOT_COM;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.net.URL;
-import java.util.stream.Stream;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 
 class RemoteAPITest {
 
-    private static Stream<Arguments> buildNewRemoteAPI_withSpecificOrDefaultSettings_verifyAPISettings() {
-        return Stream.of(
-                Arguments.of(new RemoteAPI.Builder().protocol("ftp").build(), "ftp", THE_TVDB_DOT_COM, HTTPS_DEFAULT),
-                Arguments.of(new RemoteAPI.Builder().host("some.host.com")
-                        .build(), HTTPS, "some.host.com", HTTPS_DEFAULT),
-                Arguments.of(new RemoteAPI.Builder().port(8457).build(), HTTPS, THE_TVDB_DOT_COM, 8457),
-                Arguments.of(new RemoteAPI.Builder().build(), HTTPS, THE_TVDB_DOT_COM, HTTPS_DEFAULT)
-        );
-    }
-
-    @ParameterizedTest(name = "[{index}] Remote API \"{0}\" has settings [protocol={1}, host={2}, port={3}]")
-    @MethodSource
-    void buildNewRemoteAPI_withSpecificOrDefaultSettings_verifyAPISettings(RemoteAPI remote, String protocol,
-            String host, int port) {
-        assertThat(remote.getProtocol()).isEqualTo(protocol);
-        assertThat(remote.getHost()).isEqualTo(host);
-        assertThat(remote.getPort()).isEqualTo(port);
+    @Test
+    void getDefault_verifyAPISettings() {
+        RemoteAPI remote = RemoteAPI.getDefault();
+        assertThat(remote.getProtocol()).isEqualTo(HTTPS);
+        assertThat(remote.getHost()).isEqualTo(THE_TVDB_DOT_COM);
+        assertThat(remote.getPort()).isEqualTo(HTTPS_DEFAULT);
+        assertThat(remote.getPath()).contains(API_V4);
     }
 
     @Test